+++ src/cooperation/block_tracker.h
@@ -48,7 +48,7 @@ namespace covered
 
         // (4) Get size for capacity check
 
-        uint32_t getSizeForCapacity() const;
+        uint64_t getSizeForCapacity() const;
     private:
         typedef std::unordered_map<Key, bool, KeyHasher> perkey_writeflag_t;
         typedef std::unordered_map<Key, std::unordered_set<NetworkAddr, NetworkAddrHasher>, KeyHasher> perkey_edge_blocklist_t;
