+++ src/cooperation/directory/directory_info.c
@@ -53,7 +53,7 @@ namespace covered
         return size - position;
     }
 
-    uint32_t DirectoryInfo::getSizeForCapacity() const
+    uint64_t DirectoryInfo::getSizeForCapacity() const
     {
         return sizeof(uint32_t);
     }
